generate coldstart depending on tidal spinup phase

+++ coupledmodeldriver/generate/adcirc/generate.py
@@ -1,4 +1,5 @@
 from copy import deepcopy
+from enum import Enum
 import logging
 import os
 from os import PathLike
@@ -16,6 +17,11 @@ from coupledmodeldriver.script import EnsembleCleanupScript, \
     EnsembleRunScript
 from coupledmodeldriver.utilities import LOGGER, create_symlink, \
     get_logger
+
+
+class RunPhase(Enum):
+    COLDSTART = 'coldstart'
+    HOTSTART = 'hotstart'
 
 
 def generate_adcirc_configuration(
@@ -64,7 +70,9 @@ def generate_adcirc_configuration(
 
     if use_nems:
         LOGGER.debug(f'generating NEMS configuration')
-        ensemble_configuration = NEMSADCIRCRunConfiguration.read_directory(configuration_directory)
+        ensemble_configuration = NEMSADCIRCRunConfiguration.read_directory(
+            configuration_directory
+        )
     else:
         LOGGER.debug(f'generating ADCIRC-only configuration')
         ensemble_configuration = ADCIRCRunConfiguration.read_directory(configuration_directory)
@@ -86,11 +94,11 @@ def generate_adcirc_configuration(
     use_original_mesh = ensemble_configuration['adcirc']['use_original_mesh']
 
     if use_nems:
-        nems_modeling_system = ensemble_configuration['nems'].nemspy_modeling_system
-        run_processors = nems_modeling_system.processors
+        nems_configuration = ensemble_configuration['nems'].nemspy_modeling_system
+        run_processors = nems_configuration.processors
         run_executable = ensemble_configuration['nems']['executable_path']
     else:
-        nems_modeling_system = None
+        nems_configuration = None
         run_processors = adcirc_processors
         run_executable = adcirc_executable_path
 
@@ -100,35 +108,56 @@ def generate_adcirc_configuration(
     if original_fort14_filename is None or not original_fort14_filename.exists():
         raise FileNotFoundError(f'mesh XY not found at "{original_fort14_filename}"')
 
-    if use_nems:
-        if tidal_spinup_duration is not None:
-            LOGGER.debug(f'setting spinup to {tidal_spinup_duration}')
-            nems_spinup_modeling_system = ModelingSystem(
-                nems_modeling_system.start_time - tidal_spinup_duration,
-                nems_modeling_system.start_time,
-                nems_modeling_system.interval,
-                ocn=deepcopy(nems_modeling_system['OCN']),
-                **nems_modeling_system.attributes,
-            )
-        else:
-            nems_spinup_modeling_system = None
+    adcprep_run_name = 'ADCIRC_MESH_PREP'
+    adcprep_job_script_filename = output_directory / f'job_adcprep_{platform.name.lower()}.job'
 
-        if nems_spinup_modeling_system is not None:
-            coldstart_filenames = nems_spinup_modeling_system.write(
+    coldstart_run_name = 'ADCIRC_COLDSTART'
+    coldstart_run_script_filename = (
+        output_directory / f'job_adcirc_{platform.name.lower()}.job.coldstart'
+    )
+
+    local_fort13_filename = output_directory / 'fort.13'
+    local_fort14_filename = output_directory / 'fort.14'
+    local_fort15_filename = output_directory / 'fort.15'
+
+    if tidal_spinup_duration is not None:
+        run_phase = RunPhase.HOTSTART
+        hotstart_run_name = 'ADCIRC_HOTSTART'
+        hotstart_run_script_filename = (
+            output_directory / f'job_adcirc_{platform.name.lower()}.job.hotstart'
+        )
+        run_script_filename = hotstart_run_script_filename
+    else:
+        run_phase = RunPhase.COLDSTART
+        hotstart_run_name = coldstart_run_name
+        hotstart_run_script_filename = coldstart_run_script_filename
+        run_script_filename = coldstart_run_script_filename
+
+    if use_nems:
+        if run_phase == RunPhase.HOTSTART:
+            LOGGER.debug(f'setting spinup to {tidal_spinup_duration}')
+            tidal_spinup_nems_configuration = ModelingSystem(
+                nems_configuration.start_time - tidal_spinup_duration,
+                nems_configuration.start_time,
+                nems_configuration.interval,
+                ocn=deepcopy(nems_configuration['OCN']),
+                **nems_configuration.attributes,
+            )
+            coldstart_filenames = tidal_spinup_nems_configuration.write(
                 output_directory,
                 overwrite=overwrite,
                 include_version=True,
                 create_atm_namelist_rc=False,
             )
         else:
-            coldstart_filenames = nems_modeling_system.write(
+            coldstart_filenames = nems_configuration.write(
                 output_directory,
                 overwrite=overwrite,
                 include_version=True,
                 create_atm_namelist_rc=False,
             )
         filenames = (f'"{filename.name}"' for filename in coldstart_filenames)
-        LOGGER.info(f'writing NEMS coldstart configuration: ' f'{", ".join(filenames)}')
+        LOGGER.info(f'writing NEMS coldstart configuration: {", ".join(filenames)}')
 
         for filename in coldstart_filenames:
             coldstart_filename = Path(f'{filename}.coldstart')
@@ -143,54 +172,33 @@ def generate_adcirc_configuration(
             else:
                 filename.rename(coldstart_filename)
 
-        if nems_spinup_modeling_system is not None:
-            hotstart_filenames = nems_modeling_system.write(
+        if run_phase == RunPhase.HOTSTART:
+            hotstart_filenames = nems_configuration.write(
                 output_directory,
                 overwrite=overwrite,
                 include_version=True,
                 create_atm_namelist_rc=False,
             )
             filenames = (f'"{filename.name}"' for filename in hotstart_filenames)
-            LOGGER.info(f'writing NEMS hotstart configuration: ' f'{", ".join(filenames)}')
-        else:
-            hotstart_filenames = []
+            LOGGER.info(f'writing NEMS hotstart configuration: {", ".join(filenames)}')
 
-        for filename in hotstart_filenames:
-            hotstart_filename = Path(f'{filename}.hotstart')
-            if hotstart_filename.exists():
-                os.remove(hotstart_filename)
-            if filename.absolute().is_symlink():
-                target = filename.resolve()
-                if target.absolute() in [value.resolve() for value in hotstart_filenames]:
-                    target = Path(f'{target}.hotstart')
-                create_symlink(target, hotstart_filename, relative=True)
-                os.remove(filename)
-            else:
-                filename.rename(hotstart_filename)
-
-    coldstart_directory = output_directory / 'coldstart'
-    runs_directory = output_directory / 'runs'
-
-    for directory in [coldstart_directory, runs_directory]:
-        if not directory.exists():
-            directory.mkdir()
+            for filename in hotstart_filenames:
+                hotstart_filename = Path(f'{filename}.hotstart')
+                if hotstart_filename.exists():
+                    os.remove(hotstart_filename)
+                if filename.absolute().is_symlink():
+                    target = filename.resolve()
+                    if target.absolute() in [value.resolve() for value in hotstart_filenames]:
+                        target = Path(f'{target}.hotstart')
+                    create_symlink(target, hotstart_filename, relative=True)
+                    os.remove(filename)
+                else:
+                    filename.rename(hotstart_filename)
 
     slurm_account = platform.value['slurm_account']
 
-    adcprep_run_name = 'ADC_MESH_DECOMP'
-    adcirc_coldstart_run_name = 'ADC_COLD_RUN'
-    adcirc_hotstart_run_name = 'ADC_HOT_RUN'
-
-    adcprep_job_script_filename = output_directory / f'job_adcprep_{platform.name.lower()}.job'
-    coldstart_run_script_filename = (
-        output_directory / f'job_adcirc_{platform.name.lower()}.job.coldstart'
-    )
-    hotstart_run_script_filename = (
-        output_directory / f'job_adcirc_{platform.name.lower()}.job.hotstart'
-    )
-
-    run_script_filename = output_directory / f'run_{platform.name.lower()}.sh'
-    cleanup_script_filename = output_directory / f'cleanup.sh'
+    ensemble_run_script_filename = output_directory / f'run_{platform.name.lower()}.sh'
+    ensemble_cleanup_script_filename = output_directory / f'cleanup.sh'
 
     LOGGER.debug(f'setting mesh partitioner "{adcprep_executable_path}"')
     adcprep_script = AdcircMeshPartitionJob(
@@ -208,48 +216,46 @@ def generate_adcirc_configuration(
         source_filename=source_filename,
     )
 
-    LOGGER.debug(
-        f'writing mesh partitioning job script ' f'"{adcprep_job_script_filename.name}"'
-    )
+    LOGGER.debug(f'writing mesh partitioning job script "{adcprep_job_script_filename.name}"')
     adcprep_script.write(adcprep_job_script_filename, overwrite=overwrite)
 
     LOGGER.debug(f'setting run executable "{run_executable}"')
-    if tidal_spinup_duration is not None:
-        coldstart_run_script = AdcircRunJob(
-            platform=platform,
-            slurm_tasks=run_processors,
-            slurm_account=slurm_account,
-            slurm_duration=job_duration,
-            slurm_run_name=adcirc_coldstart_run_name,
-            executable=run_executable,
-            slurm_partition=partition,
-            slurm_email_type=email_type,
-            slurm_email_address=email_address,
-            slurm_error_filename=f'{adcirc_coldstart_run_name}.err.log',
-            slurm_log_filename=f'{adcirc_coldstart_run_name}.out.log',
-            source_filename=source_filename,
-        )
-
-        LOGGER.debug(f'writing coldstart run script "{coldstart_run_script_filename.name}"')
-        coldstart_run_script.write(coldstart_run_script_filename, overwrite=overwrite)
-
-    hotstart_run_script = AdcircRunJob(
+    coldstart_run_script = AdcircRunJob(
         platform=platform,
         slurm_tasks=run_processors,
         slurm_account=slurm_account,
         slurm_duration=job_duration,
-        slurm_run_name=adcirc_hotstart_run_name,
+        slurm_run_name=coldstart_run_name,
         executable=run_executable,
         slurm_partition=partition,
         slurm_email_type=email_type,
         slurm_email_address=email_address,
-        slurm_error_filename=f'{adcirc_hotstart_run_name}.err.log',
-        slurm_log_filename=f'{adcirc_hotstart_run_name}.out.log',
+        slurm_error_filename=f'{coldstart_run_name}.err.log',
+        slurm_log_filename=f'{coldstart_run_name}.out.log',
         source_filename=source_filename,
     )
 
-    LOGGER.debug(f'writing hotstart run script ' f'"{hotstart_run_script_filename.name}"')
-    hotstart_run_script.write(hotstart_run_script_filename, overwrite=overwrite)
+    LOGGER.debug(f'writing coldstart run script "{coldstart_run_script_filename.name}"')
+    coldstart_run_script.write(coldstart_run_script_filename, overwrite=overwrite)
+
+    if run_phase == RunPhase.HOTSTART:
+        hotstart_run_script = AdcircRunJob(
+            platform=platform,
+            slurm_tasks=run_processors,
+            slurm_account=slurm_account,
+            slurm_duration=job_duration,
+            slurm_run_name=hotstart_run_name,
+            executable=run_executable,
+            slurm_partition=partition,
+            slurm_email_type=email_type,
+            slurm_email_address=email_address,
+            slurm_error_filename=f'{hotstart_run_name}.err.log',
+            slurm_log_filename=f'{hotstart_run_name}.out.log',
+            source_filename=source_filename,
+        )
+
+        LOGGER.debug(f'writing hotstart run script "{hotstart_run_script_filename.name}"')
+        hotstart_run_script.write(hotstart_run_script_filename, overwrite=overwrite)
 
     try:
         # instantiate AdcircRun object.
@@ -259,8 +265,6 @@ def generate_adcirc_configuration(
             LOGGER.debug(f'moving out of "{configuration_directory}"')
             os.chdir(starting_directory)
 
-    local_fort13_filename = output_directory / 'fort.13'
-    local_fort14_filename = output_directory / 'fort.14'
     if use_original_mesh:
         LOGGER.info(f'using original mesh from "{original_fort14_filename}"')
         if original_fort13_filename.exists():
@@ -273,17 +277,24 @@ def generate_adcirc_configuration(
             overwrite=overwrite,
             fort13=None,
             fort14='fort.14',
-            fort15=None,
+            fort15='fort.15',
             fort22=None,
             coldstart=None,
             hotstart=None,
             driver=None,
         )
 
-    if coldstart_run_script_filename.exists():
-        LOGGER.debug(f'writing coldstart configuration to ' f'"{coldstart_directory}"')
+    if local_fort15_filename.exists():
+        os.remove(local_fort15_filename)
+
+    if run_phase == RunPhase.HOTSTART:
+        tidal_spinup_directory = output_directory / 'tidal_spinup'
+        if not tidal_spinup_directory.exists():
+            tidal_spinup_directory.mkdir(parents=True, exist_ok=True)
+
+        LOGGER.debug(f'writing tidal spinup configuration to "{tidal_spinup_directory}"')
         coldstart_driver.write(
-            coldstart_directory,
+            tidal_spinup_directory,
             overwrite=overwrite,
             fort13=None if use_original_mesh else 'fort.13',
             fort14=None,
@@ -294,47 +305,51 @@ def generate_adcirc_configuration(
         if use_original_mesh:
             if local_fort13_filename.exists():
                 create_symlink(
-                    local_fort13_filename, coldstart_directory / 'fort.13', relative=True
+                    local_fort13_filename, tidal_spinup_directory / 'fort.13', relative=True
                 )
-        create_symlink(local_fort14_filename, coldstart_directory / 'fort.14', relative=True)
         create_symlink(
-            adcprep_job_script_filename, coldstart_directory / 'adcprep.job', relative=True,
+            local_fort14_filename, tidal_spinup_directory / 'fort.14', relative=True
         )
         create_symlink(
-            coldstart_run_script_filename, coldstart_directory / 'adcirc.job', relative=True,
+            adcprep_job_script_filename, tidal_spinup_directory / 'adcprep.job', relative=True,
+        )
+        create_symlink(
+            coldstart_run_script_filename,
+            tidal_spinup_directory / 'adcirc.job',
+            relative=True,
         )
         if use_nems:
             create_symlink(
                 output_directory / 'nems.configure.coldstart',
-                coldstart_directory / 'nems.configure',
+                tidal_spinup_directory / 'nems.configure',
                 relative=True,
             )
             create_symlink(
                 output_directory / 'model_configure.coldstart',
-                coldstart_directory / 'model_configure',
+                tidal_spinup_directory / 'model_configure',
                 relative=True,
             )
             create_symlink(
                 output_directory / 'config.rc.coldstart',
-                coldstart_directory / 'config.rc',
+                tidal_spinup_directory / 'config.rc',
                 relative=True,
             )
 
     perturbations = ensemble_configuration.perturb(relative_path=3)
 
-    LOGGER.info(
-        f'generating {len(perturbations)} '
-        f'"{platform.name.lower()}" configuration(s) in "{runs_directory}"'
-    )
+    runs_directory = output_directory / 'runs'
+    if not runs_directory.exists():
+        runs_directory.mkdir(parents=True, exist_ok=True)
 
-    for run_name, hotstart_configuration in perturbations.items():
-        hotstart_directory = runs_directory / run_name
-        LOGGER.debug(f'writing hotstart configuration to "{hotstart_directory}"')
+    LOGGER.info(f'generating {len(perturbations)} run configuration(s) in "{runs_directory}"')
+    for run_name, run_configuration in perturbations.items():
+        run_directory = runs_directory / run_name
+        LOGGER.debug(f'writing run configuration to "{run_directory}"')
 
-        hotstart_driver = hotstart_configuration.adcircpy_driver
+        run_driver = run_configuration.adcircpy_driver
 
-        hotstart_driver.write(
-            hotstart_directory,
+        run_driver.write(
+            run_directory,
             overwrite=overwrite,
             fort13=None if use_original_mesh else 'fort.13',
             fort14=None,
@@ -344,38 +359,36 @@ def generate_adcirc_configuration(
         )
         if use_original_mesh:
             if local_fort13_filename.exists():
-                create_symlink(
-                    local_fort13_filename, hotstart_directory / 'fort.13', relative=True
-                )
-        create_symlink(local_fort14_filename, hotstart_directory / 'fort.14', relative=True)
+                create_symlink(local_fort13_filename, run_directory / 'fort.13', relative=True)
+        create_symlink(local_fort14_filename, run_directory / 'fort.14', relative=True)
         create_symlink(
-            adcprep_job_script_filename, hotstart_directory / 'adcprep.job', relative=True,
+            adcprep_job_script_filename, run_directory / 'adcprep.job', relative=True,
         )
         create_symlink(
-            hotstart_run_script_filename, hotstart_directory / 'adcirc.job', relative=True,
+            run_script_filename, run_directory / 'adcirc.job', relative=True,
         )
         if use_nems:
             create_symlink(
-                output_directory / 'nems.configure.hotstart',
-                hotstart_directory / 'nems.configure',
+                output_directory / f'nems.configure.{run_phase.value}',
+                run_directory / 'nems.configure',
                 relative=True,
             )
             create_symlink(
-                output_directory / 'model_configure.hotstart',
-                hotstart_directory / 'model_configure',
+                output_directory / f'model_configure.{run_phase.value}',
+                run_directory / 'model_configure',
                 relative=True,
             )
             create_symlink(
-                output_directory / 'config.rc.hotstart',
-                hotstart_directory / 'config.rc',
+                output_directory / f'config.rc.{run_phase.value}',
+                run_directory / 'config.rc',
                 relative=True,
             )
-        if tidal_spinup_duration is not None:
+        if run_phase == RunPhase.HOTSTART:
             for hotstart_filename in ['fort.67.nc', 'fort.68.nc']:
                 try:
                     create_symlink(
-                        coldstart_directory / hotstart_filename,
-                        hotstart_directory / hotstart_filename,
+                        tidal_spinup_directory / hotstart_filename,
+                        run_directory / hotstart_filename,
                         relative=True,
                     )
                 except:
@@ -384,13 +397,13 @@ def generate_adcirc_configuration(
                         'you must manually link or copy this file after coldstart completes'
                     )
 
-    LOGGER.info(f'writing ensemble run script "{run_script_filename.name}"')
-    run_script = EnsembleRunScript(platform)
-    run_script.write(run_script_filename, overwrite=overwrite)
-
     cleanup_script = EnsembleCleanupScript()
-    LOGGER.debug(f'writing cleanup script "{cleanup_script_filename.name}"')
-    cleanup_script.write(cleanup_script_filename, overwrite=overwrite)
+    LOGGER.debug(f'writing cleanup script "{ensemble_cleanup_script_filename.name}"')
+    cleanup_script.write(ensemble_cleanup_script_filename, overwrite=overwrite)
+
+    LOGGER.info(f'writing ensemble run script "{ensemble_run_script_filename.name}"')
+    run_script = EnsembleRunScript(platform)
+    run_script.write(ensemble_run_script_filename, overwrite=overwrite)
 
     if starting_directory is not None:
         LOGGER.debug(f'moving out of "{configuration_directory}"')
